pyproject.toml
[tool.poetry]
name = "scrapie"
version = "0.1.0"
description = ""
authors = ["Aditya Karnam <akarnam37@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.1"
uvicorn = "^0.16.0"
Authlib = "^0.15.5"
httpx = "^0.21.1"
itsdangerous = "^2.0.1"
toml = "^0.10.2"
requests = "^2.26.0"
psycopg2 = "^2.9.2"
SQLAlchemy = "^1.4.28"
python-dotenv = "^0.19.2"
alembic = "^1.7.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
